refactor(nav): extract DoubleClickNav from ExperimentalControls
Move double-click navigation (navigateDoubleClick + _resolveStandoff) into
doubleClickNav.js: classify what's under the cursor, compute a "good view"
desired pose (navMath, pure), resolve the standoff onto a clear non-buried pose,
and hand the teleport tween to ctx.runner.run. The orchestrator keeps
navigateDoubleClick as a public delegator (viewport.js fires it on the
nav-experimental:doubleclick event).

Behaviour-preserving: 302/302 characterization tests green.
